refactor: 테스트 용 유틸 추출
